+++ bpf/libbpf/xdp_dos.c
@@ -32,12 +32,8 @@
 #include "xdp_dos.skel.h"
 #include "xdp_dos.h"
 
-#define struct_bpf	xdp_dos_bpf
-#define _bpf__open_opts	xdp_dos_bpf__open_opts
-#define _bpf__open	xdp_dos_bpf__open
-#define _bpf__open_and_load	xdp_dos_bpf__open_and_load
-#define _bpf__load	xdp_dos_bpf__load
-#define _bpf__destroy	xdp_dos_bpf__destroy
+#define NAME xdp_dos
+#include "skel_defs.h"
 
 
 static int ifindex = -1;